+++ src/main/java/com/briup/buke/dao/AdministratorDao.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import com.briup.buke.bean.Administrator;
 import com.briup.buke.bean.Reader;
 
-public interface AdministratorDao extends JpaRepository<Administrator, Integer> {
+public interface AdministratorDao extends JpaRepository<Administrator, Long> {
 	@Query(value="select * from book_administrator c where c.username=?1",nativeQuery=true)
 	public Administrator findByUsername(String username);
 }
