+++ src/main/java/com/briup/buke/dao/CategoryDao.java
@@ -10,15 +10,15 @@ import org.springframework.transaction.annotation.Transactional;
 import com.briup.buke.bean.Category;
 import com.briup.buke.bean.CategoryPack;
 
-public interface CategoryDao extends JpaRepository<Category, Integer> {
+public interface CategoryDao extends JpaRepository<Category, Long> {
 	@Query(value="select c.id from book_category c where c.name=?1",nativeQuery=true)
-	public Integer findIdByName(String name);
+	public Long findIdByName(String name);
 	
 	@Query(value="select c.name from book_category c where c.id=?1",nativeQuery=true)
-	public String findNameById(Integer id);
+	public String findNameById(Long id);
 	
 	@Transactional
 	@Modifying
 	@Query(value="delete from book_category where id= ?1",nativeQuery=true)
-	public void deleteById(Integer id);
+	public void deleteById(Long id);
 }
