+++ src/main/java/com/briup/buke/dao/ChapterDao.java
@@ -7,10 +7,10 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.briup.buke.bean.Chapter;
 
-public interface ChapterDao extends JpaRepository<Chapter, Integer> {
+public interface ChapterDao extends JpaRepository<Chapter, Long> {
 	@Query(value="select * from book_chapter c where c.article_id=?1",nativeQuery=true)
-	public List<Chapter> findByArticleId(Integer article_id);
+	public List<Chapter> findByArticleId(Long article_id);
 	
 	@Query(value="select * from book_chapter c where c.article_id=?1 order by c.id desc limit 6",nativeQuery=true)
-	public List<Chapter> findUpdateChapter(Integer id);
+	public List<Chapter> findUpdateChapter(Long id);
 }
