+++ src/main/java/com/briup/buke/dao/LinkDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.briup.buke.bean.Link;
 
-public interface LinkDao extends JpaRepository<Link, Integer> {
+public interface LinkDao extends JpaRepository<Link, Long> {
 
 }
