+++ src/main/java/com/briup/buke/service/IChapterService.java
@@ -9,13 +9,13 @@ import com.briup.buke.bean.Chapter;
 public interface IChapterService {
 	public void saveOrUpdate(Chapter chapter)throws Exception;
 	
-	public void deleteById(Integer id)throws Exception;
+	public void deleteById(Long id)throws Exception;
 	
-	public Chapter findById(Integer id)throws Exception;
+	public Chapter findById(Long id)throws Exception;
 	
-	public List<Chapter> findByArticleId(Integer article_id);
+	public List<Chapter> findByArticleId(Long article_id);
 	
-	public List<Chapter> findUpdateChapter(Integer id);
+	public List<Chapter> findUpdateChapter(Long id);
 	
 	public List<Chapter> findAll();
 }
