+++ src/main/java/com/briup/buke/service/ICommentService.java
@@ -7,11 +7,11 @@ import com.briup.buke.bean.Comment;
 public interface ICommentService {
 	public List<Comment> findAll();
 	
-	public Comment findById(Integer id)throws Exception;
+	public Comment findById(Long id)throws Exception;
 	
 	public void saveOrUpdate(Comment comment)throws Exception;
 	
-	public void deleteById(Integer id)throws Exception;
+	public void deleteById(Long id)throws Exception;
 	
-	public List<Comment> findCommentByArticleId (Integer id);
+	public List<Comment> findCommentByArticleId (Long id);
 }
